+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2024-10-19
+
+### Fixed
+
+- TransitionExtensions:
+  - ClearTransitions Unregister callback correctly now.
+  - both SetupTransition unregister callback correctly now
+
+### Changed
+
+- TransitionExtensions:
+  - SetupTransitions with multiple propertys change to dictionary of string and object and parsed to get kvp
+- README updated
+
 ## [1.0.0] - 2024-10-19
 
 ### Added